% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{as_flextable,FLQuant-method}
\alias{as_flextable,FLQuant-method}
\alias{as_flextable,FLQuants-method}
\title{Coercing FLQuant(s) into flextable}
\usage{
\S4method{as_flextable}{FLQuant}(x)

\S4method{as_flextable}{FLQuants}(x)
}
\arguments{
\item{x}{Object to transform, of class 'FLQuant' or 'FLQuants'}
}
\description{
Coercing FLQuant(s) into flextable
}
\examples{
data(ple4)
as_flextable(catch(ple4))
as_flextable(catch.n(ple4))
as_flextable(metrics(ple4))
}
