% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{stockReport}
\alias{stockReport}
\title{Template for Rmarkdown version of a stock assessment template}
\usage{
stockReport(...)
}
\arguments{
\item{...}{Any arguments to pass to bookdown::markdown_document2}
}
\value{
A document configuration, of class rmarkdown_output_format.
}
\description{
Template for Rmarkdown version of a stock assessment template
}
